# Move ArgSet to arguments package/replace too simple functions to references