--- a/compiler/bytecode/expression_generation.go
+++ b/compiler/bytecode/expression_generation.go
@@ -187,8 +187,8 @@ func (g *Generator) compileAssignExpression(is *InstructionSet, exp *ast.AssignE
 
 func (g *Generator) compileBlockArgExpression(index int, exp *ast.CallExpression, scope *scope, table *localTable) {
 	is := &InstructionSet{}
-	is.name = fmt.Sprint(index)
-	is.isType = Block
+	is.Name = fmt.Sprint(index)
+	is.InstType = Block
 
 	for i := 0; i < len(exp.BlockArguments); i++ {
 		table.set(exp.BlockArguments[i].Value)