--- a/vm/call_object.go
+++ b/vm/call_object.go
@@ -51,7 +51,7 @@ func (co *callObject) argPosition() int {
 }
 
 func (co *callObject) assignNormalArguments(stack []*Pointer) {
-	for i, paramType := range co.method.instructionSet.paramTypes.Types {
+	for i, paramType := range co.method.instructionSet.ArgSet.Types {
 		if paramType == bytecode.NormalArg {
 			co.callFrame.insertLCL(i, 0, stack[co.argPosition()].Target)
 			co.argIndex++
@@ -104,7 +104,7 @@ func (co *callObject) assignKeywordArguments(stack []*Pointer) (err error) {
 }
 
 func (co *callObject) assignSplatArgument(stack []*Pointer, arr *ArrayObject) {
-	index := len(co.method.instructionSet.paramTypes.Types) - 1
+	index := len(co.method.instructionSet.ArgSet.Types) - 1
 
 	for co.argIndex < co.argCount {
 		arr.Elements = append(arr.Elements, stack[co.argPosition()].Target)
@@ -115,8 +115,8 @@ func (co *callObject) assignSplatArgument(stack []*Pointer, arr *ArrayObject) {
 }
 
 func (co *callObject) hasKeywordParam(name string) (index int, result bool) {
-	for paramIndex, paramType := range co.method.instructionSet.paramTypes.Types {
-		paramName := co.method.instructionSet.paramTypes.Names[paramIndex]
+	for paramIndex, paramType := range co.method.instructionSet.ArgSet.Types {
+		paramName := co.method.instructionSet.ArgSet.Names[paramIndex]
 
 		if paramName == name && (paramType == bytecode.RequiredKeywordArg || paramType == bytecode.OptionalKeywordArg) {
 			index = paramIndex
@@ -143,7 +143,7 @@ func (co *callObject) hasKeywordArgument(name string) (index int, result bool) {
 }
 
 func (co *callObject) normalParamsCount() (n int) {
-	for _, at := range co.method.instructionSet.paramTypes.Types {
+	for _, at := range co.method.instructionSet.ArgSet.Types {
 		if at == bytecode.NormalArg {
 			n++
 		}