--- a/vm/method.go
+++ b/vm/method.go
@@ -44,7 +44,7 @@ func (m *MethodObject) Value() interface{} {
 }
 
 func (m *MethodObject) paramTypes() []uint8 {
-	return m.instructionSet.paramTypes.Types()
+	return m.instructionSet.paramTypes.Types
 }
 
 func (m *MethodObject) isSplatArgIncluded() bool {