Moved `ArgSet` structs to arguments.go.

I found some too simple functions just to return the value from deep structs, so replaced them.

```
benchmarking f54dc3e582496de9dd459c3858a38ff3844c3f54
benchmarking 08a436e3cf4f70a845ea3114428b42ada0d27700
x86_64-darwin17
benchmark                              old ns/op     new ns/op     delta
BenchmarkBasicMath/add-8               3298          3165          -4.03%
BenchmarkBasicMath/subtract-8          3266          3182          -2.57%
BenchmarkBasicMath/multiply-8          3308          3205          -3.11%
BenchmarkBasicMath/divide-8            3294          3213          -2.46%
BenchmarkConcurrency/concurrency-8     8511839       7654885       -10.07%

benchmark                              old allocs     new allocs     delta
BenchmarkBasicMath/add-8               33             33             +0.00%
BenchmarkBasicMath/subtract-8          33             33             +0.00%
BenchmarkBasicMath/multiply-8          33             33             +0.00%
BenchmarkBasicMath/divide-8            33             33             +0.00%
BenchmarkConcurrency/concurrency-8     53524          53520          -0.01%

benchmark                              old bytes     new bytes     delta
BenchmarkBasicMath/add-8               1568          1568          +0.00%
BenchmarkBasicMath/subtract-8          1568          1568          +0.00%
BenchmarkBasicMath/multiply-8          1568          1568          +0.00%
BenchmarkBasicMath/divide-8            1568          1568          +0.00%
BenchmarkConcurrency/concurrency-8     2729965       2729851       -0.00%
```