@st0012 2018-08-26
@hachi8833 Thanks for experimenting this! However, the benchmark on CI tells a different story
```
benchmarking 1d893503dd9c9c1cf4fc1a847fb805a02b1e2753
benchmarking b3225a9c1d8cb609b2375959e5e979744a607495
x86_64-linux
benchmark                              old ns/op     new ns/op     delta
BenchmarkBasicMath/add-2               4023          4138          +2.86%
BenchmarkBasicMath/subtract-2          3973          4137          +4.13%
BenchmarkBasicMath/multiply-2          3998          4165          +4.18%
BenchmarkBasicMath/divide-2            3995          4151          +3.90%
BenchmarkConcurrency/concurrency-2     18729110      20308166      +8.43%
```

I tested it locally as well and the result is not stable either (sometimes it get worse).
And the reason I used functions over reference is for encapsulation so that the value can't be changed outside `bytecode` package (which doesn't make any sence). I'd prefer stay this way unless the performace has obvious and stable improvement.