@hachi8833 2018-09-02
Thank you for reviewing this!
I understood your explanation and close this PR.